ws-api/requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
gunicorn==21.2.0
eventlet==0.35.1
redis==5.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
PyJWT==2.8.0
requests==2.31.0
